final commit

+++ src/main/java/org/example/kapusta/BookController.java
@@ -1,10 +1,7 @@
 package org.example.kapusta;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 public class BookController {
@@ -17,7 +14,27 @@ public class BookController {
 
     @DeleteMapping("/books/{id}")
     public ResponseEntity<String> removeBook(@PathVariable String id) {
-        bookRepository.deleteById(id);
-        return ResponseEntity.ok("Book successfully deleted");
+        Book book = bookRepository.findById(id).orElse(null);
+        if (book == null) {
+            return ResponseEntity.notFound().build();
+        }
+        if(book.getQuantity() == 1){
+            bookRepository.deleteById(id);
+            return ResponseEntity.ok("Book successfully deleted");
+        }
+        book.setQuantity(book.getQuantity() - 1);
+        bookRepository.save(book);
+        return ResponseEntity.ok("Book quantity incremented");
+    }
+
+    @PostMapping("/books/{id}")
+    public ResponseEntity<String> addOneBook(@PathVariable String id) {
+        Book book = bookRepository.findById(id).orElse(null);
+        if (book == null) {
+            return ResponseEntity.notFound().build();
+        }
+        book.setQuantity(book.getQuantity() + 1);
+        bookRepository.save(book);
+        return ResponseEntity.ok("Book quantity incremented");
     }
 }
